massive update: reshaping the entire repo to a framework
