+++ pyandroidkx/jclass/android/content/res.py
@@ -1,0 +1,3 @@
+from jnius import autoclass
+
+Configuration = autoclass("android.content.res.Configuration")
