+++ pyandroidkx/jclass/android/net.py
@@ -1,0 +1,4 @@
+from jnius import autoclass
+
+Uri = autoclass('android.net.Uri')
+ConnectivityManager = autoclass('android.net.ConnectivityManager')
