+++ pyandroidkx/jclass/android/speech/tts.py
@@ -1,0 +1,3 @@
+from jnius import autoclass
+
+TextToSpeech = autoclass('android.speech.tts.TextToSpeech')
