+++ pyandroidkx/jclass/android/view.py
@@ -1,0 +1,5 @@
+from jnius import autoclass
+
+View = autoclass('android.view.View')
+WindowManager = autoclass("android.view.WindowManager")
+LayoutParams = autoclass("android.view.WindowManager$LayoutParams")
