+++ pyandroidkx/jclass/android/webkit.py
@@ -1,0 +1,3 @@
+from jnius import autoclass
+
+CookieManager = autoclass("android.webkit.CookieManager")
