+++ pyandroidkx/jclass/android/widget.py
@@ -1,0 +1,3 @@
+from jnius import autoclass
+
+Toast = autoclass('android.widget.Toast')
