+++ pyandroidkx/jclass/androidx/core/content/__init__.py
@@ -1,0 +1,3 @@
+from jnius import autoclass
+
+ContextCompat = autoclass("androidx.core.content.ContextCompat")
