+++ pyandroidkx/jclass/java/io.py
@@ -1,0 +1,3 @@
+from jnius import autoclass
+
+File = autoclass('java.io.File')
