+++ pyandroidkx/jclass/java/net.py
@@ -1,0 +1,3 @@
+from jnius import autoclass
+
+URLConnection = autoclass("java.net.URLConnection")
