+++ pyandroidkx/jclass/java/util.py
@@ -1,0 +1,4 @@
+from jnius import autoclass
+
+Locale = autoclass('java.util.Locale')
+ArrayList = autoclass('java.util.ArrayList')
