+++ pyandroidkx/jimplement/contact.py
@@ -1,8 +1,10 @@
-from kvdroid import activity
-from kvdroid import Phone
+from typing import Union
+
+from pyandroidkx import activity
+from pyandroidkx.jclass.android.provider import Phone
 
 
-def get_contact_details(option: str = "phone_book"):
+def get_contact_details(option: str = "phone_book") -> Union[list, dict]:
     """
     option accepts this values : "phone_book", "mobile_no", "names"
 
@@ -28,17 +30,17 @@ def get_contact_details(option: str = "phone_book"):
                     if name in phone_book:
                         phone_book[name].append(number)
                     else:
-                        phone_book.update({name: [number]})
+                        phone_book[name] = [number]
                     mobile_no_set.append(number)
         finally:
             cursor.close()
 
-        if option == "names":
-            value = list(phone_book.keys())
-        elif option == "mobile_no":
+        if option == "mobile_no":
             value = mobile_no_set
+        elif option == "names":
+            value = list(phone_book.keys())
         elif option == "phone_book":
             value = phone_book
         else:
-            raise TypeError("available options are ['names', 'mobile_no', 'phone_book'] for get_contact_details")
+            raise TypeError("available options are ['names', 'mobile_no', 'phone_book'] for `get_contact_details`")
     return value
