+++ pyandroidkx/jimplement/network.py
@@ -1,7 +1,9 @@
-from kvdroid import ConnectivityManager, activity, AndroidActivity
+from pyandroidkx.jclass.android.app import AndroidActivity
+from pyandroidkx.jclass.android.net import ConnectivityManager
+from pyandroidkx import activity
 
 
-def network_state():
+def network_status():
     con_mgr = activity.getSystemService(AndroidActivity.CONNECTIVITY_SERVICE)
     conn = con_mgr.getNetworkInfo(ConnectivityManager.TYPE_WIFI).isConnectedOrConnecting()
     try:
@@ -13,4 +15,4 @@ def network_state():
         return False
 
 
-network_state = network_state()
+network_state = network_status()
